Improve support for async wiki updates.
The wiki can now be properly updated in the background using Celery after a page
has been edited by a user. The Flask configuration is passed on to Celery, and
the background updater function is only set when running the Flask application
(not when running updates via `wk`, for instance).